Switched to client/server interface to Phenix.
This avoids the need for the user to know how to set things up to 
provide the Phenix environment when running ChimeraX from a desktop 
icon. Instead, they can start the Phenix server in a separate window, 
and connect to it from ChimeraX. Calling a phenix command without the 
server started will print a message to the log explaining how to start 
the server.